soc: oplus_project: Fix pointer to integer cast
../drivers/soc/oplus/system/oplus_project/qcom/oplus_project.c:501:12: warning: cast to smaller integer type 'uint32_t' (aka 'unsigned int') from 'void *' [-Wvoid-pointer-to-int-cast]
        switch(Ptr2UINT32(p)) {
               ^~~~~~~~~~~~~
../drivers/soc/oplus/system/oplus_project/qcom/oplus_project.c:26:26: note: expanded from macro 'Ptr2UINT32'
#define Ptr2UINT32(p)    (uint32_t)(p)
                         ^~~~~~~~~~~~~

+++ drivers/soc/oplus/system/oplus_project/qcom/oplus_project.c
@@ -23,7 +23,7 @@
 #define SMEM_PROJECT 135
 
 #define UINT2Ptr(n) (uint32_t *)(n)
-#define Ptr2UINT32(p) (uint32_t)(p)
+#define Ptr2UINT32(p) (uintptr_t)(p)
 
 #define PROJECT_VERSION (0x1)
 #define PCB_VERSION (0x2)
